Aula 247 -  Carrinho de Compras - Algoritmo Pacote - Parte 2

+++ LojaVirtual/LojaVirtual/Models/Pacote.cs
@@ -10,6 +10,6 @@ namespace LojaVirtual.Models
         public int Largura { get; set; }
         public int Altura { get; set; }
         public int Comprimento { get; set; }
-        public decimal Peso { get; set; }
+        public double Peso { get; set; }
     }
 }
